+++ desktop/packages/mullvad-vpn/src/renderer/features/port-forwarding/components/index.ts
@@ -1,5 +1,7 @@
 export * from './port-forwarding-advanced';
 export * from './port-forwarding-indicator';
+export * from './port-forwarding-notifications-setting';
+export * from './port-forwarding-notifications-switch';
 export * from './port-forwarding-setting';
 export * from './port-forwarding-status';
 export * from './port-forwarding-switch';