+++ desktop/packages/mullvad-vpn/src/renderer/features/port-forwarding/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './use-nat-pmp-port-block';
 export * from './use-port-forwarding';
+export * from './use-port-forwarding-notifications';